Произвёл очистку кода от лишних комментариев
