+++ MovieQuiz/Presentation/AlertPresenter.swift
@@ -1,9 +1,3 @@
-//
-//  AlertPresenter.swift
-//  MovieQuiz
-//
-//  Created by Для учёбы   on 07/05/2024.
-//
 
 import Foundation
 import UIKit
