+++ MovieQuiz/Services/MoviesLoader.swift
@@ -1,9 +1,3 @@
-//
-//  MoviesLoader.swift
-//  MovieQuiz
-//
-//  Created by Для учёбы   on 16/05/2024.
-//
 
 import Foundation
