+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,9 +1,3 @@
-//
-//  QuestionFactoryProtocol.swift
-//  MovieQuiz
-//
-//  Created by Для учёбы   on 26/04/2024.
-//
 
 import Foundation
