+++ MovieQuiz/Services/StatisticServiceProtocol.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+
+protocol StatisticService {
+    var totalAccuracy: Double { get }
+    var gamesCount: Int { get }
+    var bestGame: GameRecord { get }
+
+    func store(correct count: Int, total amount: Int)
+}
